requirements.txt
slack-sdk==3.23.0
python-dotenv==1.0.0
requests==2.31.0
dateparser==1.2.0
python-dateutil==2.8.2
pyyaml==6.0.1
